Move `MastodonClient` to `network.client` package
