+++ core/mastodon/src/main/java/br/com/orcinus/orca/core/mastodon/feed/profile/post/MastodonDeletablePost.kt
@@ -17,9 +17,9 @@ package br.com.orcinus.orca.core.mastodon.feed.profile.post
 
 import br.com.orcinus.orca.core.feed.profile.post.DeletablePost
 import br.com.orcinus.orca.core.mastodon.MastodonCoreModule
-import br.com.orcinus.orca.core.mastodon.client.authenticateAndDelete
 import br.com.orcinus.orca.core.mastodon.instance.SomeMastodonInstance
 import br.com.orcinus.orca.core.mastodon.instanceProvider
+import br.com.orcinus.orca.core.mastodon.network.client.authenticateAndDelete
 import br.com.orcinus.orca.std.injector.Injector
 
 /**
