+++ core/mastodon/src/main/java/br/com/orcinus/orca/core/mastodon/network/client/Logger.kt
@@ -13,7 +13,7 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.client
+package br.com.orcinus.orca.core.mastodon.network.client
 
 import android.util.Log
