+++ core/mastodon/src/test/java/br/com/orcinus/orca/core/mastodon/instance/TestMastodonInstance.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2023-2024 Orcinus
+ * Copyright © 2023–2024 Orcinus
  *
  * This program is free software: you can redistribute it and/or modify it under the terms of the
  * GNU General Public License as published by the Free Software Foundation, either version 3 of the
@@ -13,13 +13,13 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.client.test.instance
+package br.com.orcinus.orca.core.mastodon.instance
 
 import br.com.orcinus.orca.core.auth.AuthenticationLock
 import br.com.orcinus.orca.core.instance.Instance
-import br.com.orcinus.orca.core.mastodon.client.Logger
-import br.com.orcinus.orca.core.mastodon.client.MastodonClient
-import br.com.orcinus.orca.core.mastodon.instance.MastodonInstance
+import br.com.orcinus.orca.core.mastodon.network.client.Logger
+import br.com.orcinus.orca.core.mastodon.network.client.MastodonClient
+import br.com.orcinus.orca.core.mastodon.network.client.test
 import br.com.orcinus.orca.core.sample.test.instance.sample
 import br.com.orcinus.orca.core.test.TestAuthenticationLock
 import br.com.orcinus.orca.core.test.TestAuthenticator
