+++ core/mastodon/src/test/java/br/com/orcinus/orca/core/mastodon/network/client/AndroidLoggerTests.kt
@@ -13,18 +13,30 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.core.mastodon.client.test.instance
+package br.com.orcinus.orca.core.mastodon.network.client
 
-import br.com.orcinus.orca.core.mastodon.client.Logger
+import android.util.Log
+import io.mockk.mockkStatic
+import io.mockk.verify
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
 
-/** [Logger] returned by [test]. */
-private val testLogger =
-  object : Logger() {
-    override fun onInfo(info: String) {}
-
-    override fun onError(error: String) {}
+@RunWith(RobolectricTestRunner::class)
+internal class AndroidLoggerTests {
+  @Test
+  fun infoCallsAndroidLogI() {
+    mockkStatic(Log::class) {
+      Logger.android.info("😮")
+      verify { Log.i(Logger.ANDROID_LOGGER_TAG, "😮") }
+    }
   }
 
-/** A no-op [Logger]. */
-internal val Logger.Companion.test
-  get() = testLogger
+  @Test
+  fun errorCallsAndroidLogE() {
+    mockkStatic(Log::class) {
+      Logger.android.error("😵")
+      verify { Log.e(Logger.ANDROID_LOGGER_TAG, "😵") }
+    }
+  }
+}
